+++ vue/rules/rules.vue.js
@@ -1,0 +1,11 @@
+
+const vue_rules = {
+    test: /\.vue?$/,
+    use: [
+        {
+            loader: 'vue-loader',
+        }
+    ]
+}
+
+module.exports = vue_rules